server/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
snowflake-connector-python==3.16.0
python-dotenv==1.1.1 